requirements.txt
asgiref==3.4.1
attrs==21.2.0
Django==3.2.5
django-cors-headers==3.7.0
django-extensions==3.1.3
djangorestframework==3.12.4
psycopg2==2.9.1
pytz==2021.1
sqliteparser==0.2.5
sqlparse==0.4.1
